refactor: collapse duplicated mac/iOS brand palette and Deepgram TTS transport into SpeakCore
- Move the brand color palette to SpeakCore (single SwiftUI source for both
  platforms); delete the SpeakApp and SpeakiOS copies and add the missing
  SpeakCore imports where mac views used the palette.
- Extract the shared Deepgram Aura HTTP transport (request construction,
  auth header, response classification, key validation, pricing constant)
  into SpeakCore.DeepgramTTSAPI (Foundation/URLSession only, Sendable).
  The macOS actor keeps keychain lookup, file handling, duration and cost;
  the iOS client keeps AVAudioSession playback. Error mapping preserves each
  platform's existing error types and messages, and the API key still
  travels only in the Authorization header.
- Merge the unique assertions from the duplicate SpeakAppTests
  SensitiveHeaderRedactor suite (exact redaction format, whitespace
  trimming, short-value sensitive keys, debug-snapshot integration) into
  the SpeakCoreTests suite and delete the duplicate file.
