--- Sources/SpeakApp/BrandColors.swift
@@ -1,9 +1,0 @@
-import SwiftUI
-
-extension Color {
-  static let brandAccent = Color(red: 1.0, green: 0.42, blue: 0.24)
-  static let brandAccentWarm = Color(red: 1.0, green: 0.61, blue: 0.29)
-  static let brandAccentDeep = Color(red: 0.89, green: 0.32, blue: 0.18)
-  static let brandLagoon = Color(red: 0.18, green: 0.83, blue: 0.75)
-  static let brandLagoonDeep = Color(red: 0.10, green: 0.67, blue: 0.60)
-}
