+++ Sources/SpeakApp/Views/Settings/PronunciationDictionaryView.swift
@@ -1,3 +1,4 @@
+import SpeakCore
 import SwiftUI
 import UniformTypeIdentifiers
